pyproject.toml
[tool.poetry]
name = "sanipro"
version = "0.1.0"
description = "Sanitizes prompts."
authors = ["Yokkin"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.10"
typing-extensions = "^4.12.2"
networkx = "^3.4.2"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
